+++ src/models/Plant.java
@@ -8,7 +8,7 @@ public abstract class Plant extends GraphicalRepresentative {
     public boolean isEdible;
     protected double kcal = 0;
 
-    public Plant(final Simulation simulation, final int x, final int y, final String path) {
+    public Plant(final Simulation simulation, final double x, final double y, final String path) {
         super(simulation, x, y, Integer.parseInt(Config.get("plants_pixel_width")), Integer.parseInt(Config.get("plants_pixel_height")), path);
     }
